deleting catalog items done

+++ back/catalogue/src/main/java/com/example/catalogue/controller/VehicleFuelTypeController.java
@@ -5,10 +5,12 @@ import com.example.catalogue.model.VehicleFuelType;
 import com.example.catalogue.service.VehicleFuelTypeService;
 import org.aspectj.weaver.ast.Not;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
 
@@ -17,6 +19,9 @@ import java.util.List;
 public class VehicleFuelTypeController {
     @Autowired
     private VehicleFuelTypeService vehicleFuelTypeService;
+
+    @Autowired
+    private RestTemplate restTemplate;
 
     /**
      * GET server/catalogue/vehicleFuelType/{id}
@@ -40,6 +45,13 @@ public class VehicleFuelTypeController {
     @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Notification> deleteVehicleFuelType(@PathVariable String id) {
         try {
+            ResponseEntity<List> response = restTemplate
+                    .exchange("http://vehicle/search/fuelType/" + id, HttpMethod.GET, null, List.class);
+            List<Long> vehicleList = response.getBody();
+            if(vehicleList.size() != 0) {
+                return new ResponseEntity<>(new Notification("There is a vehicle registered with fuel type id " + id + "\nFuel type wasn't deleted.", false), HttpStatus.CONFLICT);
+            }
+
             vehicleFuelTypeService.deleteOneFuelType(id);
             return new ResponseEntity<>(new Notification("Successfully deleted fuel type id = " + id, true), HttpStatus.OK);
         } catch (Exception e) {
