+++ back/catalogue/src/main/java/com/example/catalogue/controller/VehicleMakeController.java
@@ -5,10 +5,12 @@ import com.example.catalogue.model.VehicleMake;
 import com.example.catalogue.repository.VehicleMakeRepository;
 import com.example.catalogue.service.VehicleMakeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
 
@@ -17,6 +19,9 @@ import java.util.List;
 public class VehicleMakeController {
     @Autowired
     private VehicleMakeService vehicleMakeService;
+
+    @Autowired
+    private RestTemplate restTemplate;
 
 //    /**
 //     * GET server/catalogue/vehicleMake/byModel/{id}
@@ -54,6 +59,13 @@ public class VehicleMakeController {
     @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Notification> deleteVehicleMake(@PathVariable String id) {
         try {
+            ResponseEntity<List> response = restTemplate
+                    .exchange("http://vehicle/search/make/" + id, HttpMethod.GET, null, List.class);
+            List<Long> vehicleList = response.getBody();
+            if(vehicleList.size() != 0) {
+                return new ResponseEntity<>(new Notification("There is a vehicle registered with make id " + id + "\nMake wasn't deleted.", false), HttpStatus.CONFLICT);
+            }
+
             vehicleMakeService.deleteOneMake(id);
             return new ResponseEntity<>(new Notification("Successfully deleted vehicle make id = " + id, true), HttpStatus.OK);
         } catch (Exception e) {
