+++ back/catalogue/src/main/java/com/example/catalogue/controller/VehicleModelController.java
@@ -5,10 +5,13 @@ import com.example.catalogue.model.VehicleModel;
 import com.example.catalogue.service.VehicleMakeService;
 import com.example.catalogue.service.VehicleModelService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
 
@@ -21,6 +24,8 @@ public class VehicleModelController {
     @Autowired
     private VehicleMakeService vehicleMakeService;
 
+    @Autowired
+    private RestTemplate restTemplate;
 
     /**
      * GET server/catalogue/vehicleModel/{id}
@@ -53,6 +58,12 @@ public class VehicleModelController {
     @DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Notification> deleteVehicleModel(@PathVariable String id) {
         try {
+            ResponseEntity<List> response = restTemplate
+                    .exchange("http://vehicle/search/model/" + id, HttpMethod.GET, null, List.class);
+            List<Long> vehicleList = response.getBody();
+            if(vehicleList.size() != 0) {
+                return new ResponseEntity<>(new Notification("There is a vehicle registered with model id " + id + "\nModel wasn't deleted.", false), HttpStatus.CONFLICT);
+            }
             vehicleModelService.deleteOneModel(id);
             return new ResponseEntity<>(new Notification("Successfully deleted vehicle model id = " + id, true), HttpStatus.OK);
         } catch (Exception e) {
